+++ composeApp/src/main/kotlin/com/example/a62550_foodapp/db/entity/RecipeItem.kt
@@ -26,10 +26,10 @@ import androidx.room.Index
 )
 data class RecipeItem(
     @ColumnInfo(name = "recipe_id")
-    val recipeId: Int,
+    val recipeId: Long,
 
     @ColumnInfo(name = "item_group_id")
-    val itemGroupId: Int,
+    val itemGroupId: Long,
 
     @ColumnInfo(name = "size_of_one_portion")
     val sizeOfOnePortion: Int
